% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fermentR.R
\docType{package}
\name{fermentR}
\alias{fermentR}
\alias{fermentR-package}
\title{fermentR: A package of converters and calculators for brewers of beer and
other fermented drinks.}
\arguments{
\item{Adjunct}{z}

\item{Grist}{The mix of grains and adjuncts which will be combined with liquor in the mash.}

\item{Gyle}{z}

\item{Liquor}{Water which will be used during the brewing process.}

\item{Malt}{y}

\item{Mash}{x}

\item{Sparge}{y}

\item{Wort}{x}
}
\description{
This is a placeholder description
}
\details{
The fermentR package covers areas of interest to brewers:

Converting between units of gravity (which measure concentration of sugars in water).
Calculating useful characteristics water chemistry and suggesting treatment (planned).
Calculating composition and temperature of mashes.
Calculating usage and effect of hops.
Calculating amounts of alcohol in fermented worts.
Calculating gas volumes in fermented worts and suggesting gas line lengths to control dispense (planned).
}
\section{some more text}{

 in this section...
}

