% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashcalcs.R
\name{mashtemp}
\alias{mashtemp}
\title{Calculate mash composition when brewing beer}
\usage{
mashtemp(ww, wg, tm, tw, tg, shw = 1, shg = 0.4)
}
\arguments{
\item{ww}{weight of water in Kg}

\item{wg}{weight of grain in Kg}

\item{tm}{mash temperature in degrees C}

\item{tw}{temperature of water in C before mixing}

\item{tg}{temperature of grain in C before mixing}

\item{shw}{specific heat of water in kJ/KgC. Defaults to 1.0kJ/KgC}

\item{shg}{specific heat of grain in kJ/KgC. Defaults to 0.4kJ/KgC}
}
\value{
the missing variable in the equation tm=(shw*tw*ww+shg*tg*wg)/(shw*ww+shg*wg)
}
\description{
Calculate mash composition when brewing beer
}
\examples{
# calculate temperature of water when mashing in
mashtemp(ww = 5, wg = 5, tm = NA, tw = 80, tg = 20)
# calculate strike temperature
mashtemp(5,5,63,NA,20)
}

