% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashcalcs.R
\name{strike}
\alias{strike}
\title{Strike Temperature calculator}
\usage{
strike(ww, wg, tm, tg)
}
\arguments{
\item{ww}{weight of water in Kg}

\item{wg}{weight of grain in Kg}

\item{tm}{mash temperature in degrees C}

\item{tg}{temperature of grain in C before mixing}
}
\value{
the strike temperature in degreed C
}
\description{
Calculate the strike temperature for hot liquor to achieve a given mash temperature
}
\examples{
# calculate strike temperature
strike(5,5,63,20)
}

