% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.R
\name{gravity}
\alias{gravity}
\title{Convert between Brix, Plato and Specific Gravity in unfermened worts}
\usage{
gravity(x, from = "brix", to = "sg")
}
\arguments{
\item{x}{a numeric vector giving the measurement value in its original units.}

\item{from}{the unit in which the measurement was made.}

\item{to}{the unit to which the measurement is to be converted.}
}
\value{
numeric vector in the unit specified
}
\description{
Convert between Brix, Plato and Specific Gravity in unfermened worts
}

