% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hop_calcs.R
\name{hop_calcs}
\alias{hop_calcs}
\title{Hop utilization calculations}
\usage{
hop_calcs(hw, wv, ibu, aa, utilization)
}
\arguments{
\item{hw}{Weight of hops in g}

\item{wv}{Wort volume in L}

\item{ibu}{International Bittering Units in mg/L}

\item{aa}{Percentage Alpha Acids}

\item{utilization}{Percentage hop utilization}
}
\value{
the missing variable in the equation hw = (wv*ibu)/(aa*utilization*1000)
}
\description{
Calculate hop requirements for brewing beer
}
\examples{
# calculate the weight of hops to achieve 25 IBU using Cascade at 10.3\% Alpha Acids and 30 \% utilization
hop_calcs(hw = NA, wv = 21, ibu = 25, aa = 0.103, utilization = 0.30)
}
\references{
Lewis MJ, Young, TW. (2001). Brewing, 2nd Edition. Klewer Academic / Plenum Publishers
}

