% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_calcs.R
\name{mash_calcs}
\alias{mash_calcs}
\alias{step_mash_by_boiling}
\alias{strike}
\title{Mash temperature calculator}
\usage{
mash_calcs(ml = NA, mg = NA, tm = NA, tl = NA, tg = NA, shl = 1,
  shg = 0.4)

strike(ml, mg, tm, tg)

step_mash_by_boiling(ml, mg, tm, tg)
}
\arguments{
\item{ml}{mass of liquor in Kg (or volume in L)}

\item{mg}{mass of grist in Kg}

\item{tm}{temperature of mash in degrees C after mixing}

\item{tl}{temperature of liquor in degrees C before mixing}

\item{tg}{temperature of grist in degrees C before mixing}

\item{shl}{specific heat of liquor relative to pure water. Defaults to 1.0}

\item{shg}{specific heat of grist relative to the liquor. Defaults to 0.4}

\item{shw}{specific heat of pure water. This is 4.181kJ/Kg/K}
}
\value{
the missing variable in the equation tm=(shl*tl*ml+shg*tg*mg)/(shl*ml+shg*mg)
}
\description{
Use \code{mash_calc} to calculate mash composition when brewing beer. Use \code{strike}
to calculate the strike temperature for hot liquor to achieve a given mash
temperature. The word liquor is used here in the brewing sense to mean water
which is mixed into the mash.
}
\examples{
# calculate temperature of liquor when mashing in
mash_calcs(ml = 5, mg = 5, tm = NA, tl = 80, tg = 20)
# calculate strike temperature
mash_calcs(5,5,63,NA,20)
# calculate strike temperature
strike(5,5,63,20)
}
\references{
Lewis MJ, Young, TW. (2001). Brewing, 2nd Edition. Klewer Academic / Plenum Publishers

B.K Bala & J.L Woods (1991) "Physical and Thermal Properties of Malt", Drying
 Technology, 9:4, 1091-1104, DOI: 10.1080/07373939108916735

"Feel the Mash Heat", BYO Magazine, Sept 1997. http://byo.com/stories/item/627-feel-the-mash-heat
}

